refactor: split nation, settlement, and facility logic out of WorldGenerator
WorldGenerator was 6,500+ lines mixing physical geography with political,
population, and economic setup. Split into focused modules:

- NationGenerator (nation_generator.h/.cpp): form_nations, seed_nomadic_population,
  seed_nation_capitals + terrain resistance, language family, and border change helpers.
  ~770 lines covering Stages 9.5-9.7.

- SettlementGenerator (settlement_generator.h/.cpp): seed_population_attractiveness,
  create_npcs, create_businesses. ~430 lines covering Stage 9 and NPC/business seeding.

- FacilityGenerator (facility_generator.h/.cpp): create_facilities, seed_technology +
  facility_categories_for_sector helper. ~220 lines.

WorldGenerator retains physical geography (Stages 1-8), province setup, commentary
(Stage 10), and output (Stage 11). Orchestration in generate() calls new classes.
No public API changes — all tests pass unchanged (1110 tests, 261512 assertions).
