+++ simulation/core/CMakeLists.txt
@@ -12,6 +12,9 @@ add_library(econlife_core STATIC
     world_gen/facility_type_catalog.cpp
     world_gen/technology_catalog.cpp
     world_gen/world_generator.cpp
+    world_gen/nation_generator.cpp
+    world_gen/settlement_generator.cpp
+    world_gen/facility_generator.cpp
     world_gen/h3_utils.cpp
     config/package_config.cpp
 )
