+++ simulation/core/world_gen/world_generator.cpp
@@ -17,7 +17,10 @@
 
 #include <nlohmann/json.hpp>
 
+#include "core/world_gen/facility_generator.h"
 #include "core/world_gen/h3_utils.h"
+#include "core/world_gen/nation_generator.h"
+#include "core/world_gen/settlement_generator.h"
 
 namespace econlife {
 
@@ -120,10 +123,10 @@ WorldState WorldGenerator::generate(const WorldGeneratorConfig& config) {
     create_markets(world, rng, catalog, config);
 
     // Step 5: NPC population
-    create_npcs(world, rng, config);
+    SettlementGenerator::create_npcs(world, rng, config);
 
     // Step 6: Businesses
-    create_businesses(world, rng, config);
+    SettlementGenerator::create_businesses(world, rng, config);
 
     // Step 7: Facilities (assign recipes to businesses)
     RecipeCatalog recipe_catalog;
@@ -135,7 +138,7 @@ WorldState WorldGenerator::generate(const WorldGeneratorConfig& config) {
         facility_type_catalog.load_from_csv(config.facility_types_filepath);
     }
     if (recipe_catalog.size() > 0 && facility_type_catalog.size() > 0) {
-        create_facilities(world, rng, recipe_catalog, facility_type_catalog, config);
+        FacilityGenerator::create_facilities(world, rng, recipe_catalog, facility_type_catalog, config);
     }
 
     // Store loaded recipes in WorldState for production module access.
@@ -162,21 +165,21 @@ WorldState WorldGenerator::generate(const WorldGeneratorConfig& config) {
     // Step 7b: Stage 9 — Population attractiveness (WorldGen v0.18).
     // Adjusts total_population from geology/climate attractiveness score (bounded ±40%).
     // Runs after detect_special_features() so permafrost and island flags are set.
-    seed_population_attractiveness(world, rng, config);
+    SettlementGenerator::seed_population_attractiveness(world, rng, config);
 
     // Step 7b2: Stage 9.5 — Nation formation (WorldGen v0.18).
     // Replaces the single placeholder nation with physics-derived multi-nation assignment.
     // Seeds nation territories via Voronoi growth, language families, border changes.
     // Runs after seed_population_attractiveness() (reads settlement_attractiveness).
-    form_nations(world, rng, config);
+    NationGenerator::form_nations(world, rng, config);
 
     // Step 7b3: Stage 9.6 — Nomadic population (WorldGen v0.18).
     // Seeds pastoral_carrying_capacity and nomadic_population_fraction from climate.
-    seed_nomadic_population(world, rng, config);
+    NationGenerator::seed_nomadic_population(world, rng, config);
 
     // Step 7b4: Stage 9.7 — Nation capital seeding (WorldGen v0.18).
     // Selects highest settlement_attractiveness province per nation as capital.
-    seed_nation_capitals(world, config);
+    NationGenerator::seed_nation_capitals(world, config);
 
     // Step 7c: Stage 10 — World commentary pipeline (WorldGen v0.18).
     // commentary_depth controls how much text is generated:
@@ -225,7 +228,7 @@ WorldState WorldGenerator::generate(const WorldGeneratorConfig& config) {
         tech_catalog.load_nodes_csv(nodes_path);
         tech_catalog.load_ceilings_csv(ceilings_path);
     }
-    seed_technology(world, rng, tech_catalog, config);
+    FacilityGenerator::seed_technology(world, rng, tech_catalog, config);
 
     // Step 9: Stage 11 — Output world.json if path configured.
     if (!config.output_world_file.empty()) {
@@ -789,529 +792,6 @@ void WorldGenerator::create_markets(WorldState& world, DeterministicRNG& rng,
             world.regional_markets.push_back(m);
             world.provinces[p].market_ids.push_back(
                 static_cast<uint32_t>(world.regional_markets.size()) - 1);
-        }
-    }
-}
-
-// ===========================================================================
-// NPC population seeding
-// ===========================================================================
-
-void WorldGenerator::create_npcs(WorldState& world, DeterministicRNG& rng,
-                                 const WorldGeneratorConfig& config) {
-    // Role distribution weights — determines the workforce composition.
-    // Workers are the majority; specialists are rarer.
-    struct RoleWeight {
-        NPCRole role;
-        uint32_t weight;  // relative frequency
-    };
-    static constexpr RoleWeight role_distribution[] = {
-        {NPCRole::worker, 40},
-        {NPCRole::corporate_executive, 5},
-        {NPCRole::middle_manager, 8},
-        {NPCRole::banker, 3},
-        {NPCRole::lawyer, 3},
-        {NPCRole::accountant, 3},
-        {NPCRole::journalist, 2},
-        {NPCRole::politician, 3},
-        {NPCRole::community_leader, 3},
-        {NPCRole::law_enforcement, 5},
-        {NPCRole::regulator, 2},
-        {NPCRole::prosecutor, 1},
-        {NPCRole::judge, 1},
-        {NPCRole::criminal_operator, 4},
-        {NPCRole::criminal_enforcer, 3},
-        {NPCRole::fixer, 2},
-        {NPCRole::union_organizer, 1},
-        {NPCRole::family_member, 8},
-        {NPCRole::media_editor, 1},
-        {NPCRole::ngo_investigator, 1},
-    };
-    static constexpr uint32_t total_weight =
-        40 + 5 + 8 + 3 + 3 + 3 + 2 + 3 + 3 + 5 + 2 + 1 + 1 + 4 + 3 + 2 + 1 + 8 + 1 + 1;
-
-    // Distribute NPCs across provinces weighted by population.
-    uint64_t total_pop = 0;
-    for (const auto& p : world.provinces) {
-        total_pop += p.demographics.total_population;
-    }
-    if (total_pop == 0)
-        total_pop = 1;
-
-    uint32_t npc_id_counter = 100;  // start NPC IDs at 100
-
-    for (uint32_t p = 0; p < static_cast<uint32_t>(world.provinces.size()); ++p) {
-        // NPCs per province proportional to population.
-        float pop_fraction = static_cast<float>(world.provinces[p].demographics.total_population) /
-                             static_cast<float>(total_pop);
-        uint32_t npcs_for_province =
-            std::max(1u, static_cast<uint32_t>(config.npc_count * pop_fraction + 0.5f));
-
-        for (uint32_t i = 0; i < npcs_for_province; ++i) {
-            // Select role via weighted random.
-            uint32_t roll = rng.next_uint(total_weight);
-            NPCRole role = NPCRole::worker;  // default
-            uint32_t cumulative = 0;
-            for (const auto& rw : role_distribution) {
-                cumulative += rw.weight;
-                if (roll < cumulative) {
-                    role = rw.role;
-                    break;
-                }
-            }
-
-            NPC npc{};
-            npc.id = npc_id_counter++;
-            npc.role = role;
-            npc.home_province_id = p;
-            npc.current_province_id = p;
-            npc.travel_status = static_cast<NPCTravelStatus>(0);  // resident
-            npc.status = NPCStatus::active;
-
-            // Capital varies by role.
-            float capital_base;
-            switch (role) {
-                case NPCRole::corporate_executive:
-                case NPCRole::banker:
-                    capital_base = 50000.0f + static_cast<float>(rng.next_uint(100000));
-                    break;
-                case NPCRole::lawyer:
-                case NPCRole::politician:
-                case NPCRole::judge:
-                case NPCRole::media_editor:
-                    capital_base = 30000.0f + static_cast<float>(rng.next_uint(50000));
-                    break;
-                case NPCRole::middle_manager:
-                case NPCRole::accountant:
-                    capital_base = 20000.0f + static_cast<float>(rng.next_uint(30000));
-                    break;
-                case NPCRole::criminal_operator:
-                    capital_base = 15000.0f + static_cast<float>(rng.next_uint(80000));
-                    break;
-                case NPCRole::fixer:
-                    capital_base = 10000.0f + static_cast<float>(rng.next_uint(60000));
-                    break;
-                default:
-                    capital_base = 5000.0f + static_cast<float>(rng.next_uint(25000));
-                    break;
-            }
-            npc.capital = capital_base;
-
-            // Risk tolerance varies by role.
-            switch (role) {
-                case NPCRole::criminal_operator:
-                case NPCRole::criminal_enforcer:
-                case NPCRole::fixer:
-                    npc.risk_tolerance = 0.5f + rng.next_float() * 0.4f;
-                    break;
-                case NPCRole::law_enforcement:
-                case NPCRole::judge:
-                    npc.risk_tolerance = 0.2f + rng.next_float() * 0.3f;
-                    break;
-                default:
-                    npc.risk_tolerance = 0.2f + rng.next_float() * 0.6f;
-                    break;
-            }
-
-            // Social capital varies by role.
-            switch (role) {
-                case NPCRole::community_leader:
-                case NPCRole::politician:
-                    npc.social_capital = 0.5f + rng.next_float() * 0.4f;
-                    break;
-                case NPCRole::journalist:
-                case NPCRole::media_editor:
-                    npc.social_capital = 0.3f + rng.next_float() * 0.4f;
-                    break;
-                default:
-                    npc.social_capital = 0.1f + rng.next_float() * 0.4f;
-                    break;
-            }
-
-            npc.movement_follower_count = 0;
-
-            // Motivation vector — 8 weights summing to 1.0.
-            // Order: financial_security, social_standing, personal_safety, power_influence,
-            //        ideology, loyalty, self_preservation, risk_seeking
-            float weights[8];
-            float weight_sum = 0.0f;
-            for (int w = 0; w < 8; ++w) {
-                weights[w] = 0.05f + rng.next_float() * 0.3f;
-                weight_sum += weights[w];
-            }
-            // Bias based on role.
-            switch (role) {
-                case NPCRole::corporate_executive:
-                case NPCRole::banker:
-                    weights[0] *= 2.0f;        // financial_security
-                    weight_sum += weights[0];  // adjust
-                    break;
-                case NPCRole::politician:
-                case NPCRole::community_leader:
-                    weights[3] *= 2.0f;  // power_influence
-                    weight_sum += weights[3];
-                    break;
-                case NPCRole::criminal_operator:
-                case NPCRole::criminal_enforcer:
-                    weights[0] *= 1.5f;  // financial
-                    weights[7] *= 1.5f;  // risk_seeking
-                    weight_sum += weights[0] * 0.5f + weights[7] * 0.5f;
-                    break;
-                case NPCRole::law_enforcement:
-                case NPCRole::prosecutor:
-                    weights[4] *= 1.5f;  // ideology
-                    weights[6] *= 1.5f;  // self_preservation
-                    weight_sum += weights[4] * 0.5f + weights[6] * 0.5f;
-                    break;
-                default:
-                    break;
-            }
-            // Recalculate sum and normalize to 1.0.
-            weight_sum = 0.0f;
-            for (int w = 0; w < 8; ++w)
-                weight_sum += weights[w];
-            for (int w = 0; w < 8; ++w)
-                weights[w] /= weight_sum;
-            npc.motivations.weights = {weights[0], weights[1], weights[2], weights[3],
-                                       weights[4], weights[5], weights[6], weights[7]};
-
-            world.provinces[p].significant_npc_ids.push_back(npc.id);
-            world.significant_npcs.push_back(std::move(npc));
-        }
-    }
-}
-
-// ===========================================================================
-// Business seeding
-// ===========================================================================
-
-void WorldGenerator::create_businesses(WorldState& world, DeterministicRNG& rng,
-                                       const WorldGeneratorConfig& config) {
-    // Business count: roughly 1 per 10 NPCs, minimum 1 per province.
-    uint32_t total_businesses = std::max(static_cast<uint32_t>(world.provinces.size()),
-                                         static_cast<uint32_t>(world.significant_npcs.size()) / 10);
-
-    // Sector distribution per province — influenced by archetype.
-    // Collect executive/corporate NPCs as potential owners.
-    struct OwnerCandidate {
-        uint32_t npc_id;
-        uint32_t province_id;
-    };
-    std::vector<OwnerCandidate> owner_pool;
-    for (const auto& npc : world.significant_npcs) {
-        if (npc.role == NPCRole::corporate_executive || npc.role == NPCRole::middle_manager ||
-            npc.role == NPCRole::criminal_operator || npc.role == NPCRole::fixer) {
-            owner_pool.push_back({npc.id, npc.home_province_id});
-        }
-    }
-
-    // If not enough owners, also include bankers and lawyers.
-    if (owner_pool.size() < total_businesses) {
-        for (const auto& npc : world.significant_npcs) {
-            if (npc.role == NPCRole::banker || npc.role == NPCRole::lawyer) {
-                owner_pool.push_back({npc.id, npc.home_province_id});
-            }
-        }
-    }
-
-    // Fallback: any NPC can own a business.
-    if (owner_pool.empty()) {
-        for (const auto& npc : world.significant_npcs) {
-            owner_pool.push_back({npc.id, npc.home_province_id});
-        }
-    }
-
-    // Sector options for formal businesses.
-    static constexpr BusinessSector formal_sectors[] = {
-        BusinessSector::manufacturing, BusinessSector::food_beverage,
-        BusinessSector::retail,        BusinessSector::services,
-        BusinessSector::agriculture,   BusinessSector::technology,
-        BusinessSector::real_estate,   BusinessSector::energy,
-        BusinessSector::finance,       BusinessSector::transport_logistics,
-        BusinessSector::media,         BusinessSector::research,
-    };
-    static constexpr uint32_t formal_sector_count =
-        sizeof(formal_sectors) / sizeof(formal_sectors[0]);
-
-    // Business profiles.
-    static constexpr BusinessProfile profiles[] = {
-        BusinessProfile::cost_cutter,
-        BusinessProfile::quality_player,
-        BusinessProfile::fast_expander,
-        BusinessProfile::defensive_incumbent,
-    };
-
-    uint32_t biz_id_counter = 1000;
-
-    for (uint32_t b = 0; b < total_businesses; ++b) {
-        uint32_t owner_idx = rng.next_uint(static_cast<uint32_t>(owner_pool.size()));
-        const auto& owner = owner_pool[owner_idx];
-
-        bool is_criminal = false;
-        // Criminal businesses: ~10% of total, owned by criminal operators.
-        for (const auto& npc : world.significant_npcs) {
-            if (npc.id == owner.npc_id &&
-                (npc.role == NPCRole::criminal_operator || npc.role == NPCRole::fixer)) {
-                is_criminal = (rng.next_uint(3) == 0);  // 33% chance for criminal NPCs
-                break;
-            }
-        }
-
-        NPCBusiness biz{};
-        biz.id = biz_id_counter++;
-        biz.owner_id = owner.npc_id;
-        biz.province_id = owner.province_id;
-        biz.criminal_sector = is_criminal;
-
-        if (is_criminal) {
-            biz.sector = BusinessSector::criminal;
-        } else {
-            biz.sector = formal_sectors[rng.next_uint(formal_sector_count)];
-        }
-
-        biz.profile = profiles[rng.next_uint(4)];
-
-        // Cash and revenue based on province demographics.
-        float province_wealth = 1.0f;
-        if (owner.province_id < world.provinces.size()) {
-            province_wealth =
-                world.provinces[owner.province_id].demographics.income_high_fraction * 2.0f +
-                world.provinces[owner.province_id].demographics.income_middle_fraction;
-        }
-        biz.cash = (20000.0f + static_cast<float>(rng.next_uint(200000))) * province_wealth;
-        biz.revenue_per_tick = (200.0f + static_cast<float>(rng.next_uint(800))) * province_wealth;
-        biz.cost_per_tick =
-            biz.revenue_per_tick * (0.6f + rng.next_float() * 0.3f);  // 60-90% cost ratio
-
-        biz.market_share = 0.02f + rng.next_float() * 0.15f;
-        biz.strategic_decision_tick = rng.next_uint(90);  // stagger quarterly decisions
-        biz.dispatch_day_offset = static_cast<uint8_t>(biz.id % 30);
-        biz.actor_tech_state.effective_tech_tier = 1.0f;
-        biz.regulatory_violation_severity = is_criminal ? 0.3f + rng.next_float() * 0.5f : 0.0f;
-        biz.default_activity_scope =
-            is_criminal ? VisibilityScope::concealed : VisibilityScope::institutional;
-        biz.deferred_salary_liability = 0.0f;
-        biz.accounts_payable_float = 0.0f;
-
-        world.npc_businesses.push_back(std::move(biz));
-    }
-}
-
-// ===========================================================================
-// Facility assignment — gives each business 1-3 facilities with recipes
-// ===========================================================================
-
-// Maps business sector to preferred facility type categories.
-static std::vector<std::string> facility_categories_for_sector(BusinessSector sector) {
-    switch (sector) {
-        case BusinessSector::manufacturing:
-            return {"manufacturing", "processing"};
-        case BusinessSector::food_beverage:
-            return {"manufacturing", "agriculture"};
-        case BusinessSector::agriculture:
-            return {"agriculture"};
-        case BusinessSector::energy:
-            return {"extraction", "processing"};
-        case BusinessSector::technology:
-            return {"manufacturing"};
-        case BusinessSector::research:
-            return {"manufacturing"};
-        case BusinessSector::transport_logistics:
-            return {"processing", "manufacturing"};
-        case BusinessSector::criminal:
-            return {"processing", "manufacturing", "agriculture"};
-        default:
-            return {"manufacturing", "processing"};
-    }
-}
-
-void WorldGenerator::create_facilities(WorldState& world, DeterministicRNG& rng,
-                                       const RecipeCatalog& recipes,
-                                       const FacilityTypeCatalog& facility_types,
-                                       const WorldGeneratorConfig& config) {
-    uint32_t facility_id_counter = 5000;
-
-    for (auto& biz : world.npc_businesses) {
-        // Determine how many facilities this business gets (1-3).
-        uint32_t facility_count = 1 + rng.next_uint(3);
-
-        // Get preferred categories for this business sector.
-        auto preferred_cats = facility_categories_for_sector(biz.sector);
-
-        // Collect candidate facility types from preferred categories.
-        std::vector<const FacilityType*> candidate_types;
-        for (const auto& cat : preferred_cats) {
-            auto cat_types = facility_types.by_category(cat);
-            for (const auto* ft : cat_types) {
-                candidate_types.push_back(ft);
-            }
-        }
-
-        // Fallback: all facility types if no match.
-        if (candidate_types.empty()) {
-            for (const auto& ft : facility_types.all()) {
-                candidate_types.push_back(&ft);
-            }
-        }
-        if (candidate_types.empty())
-            continue;
-
-        for (uint32_t f = 0; f < facility_count; ++f) {
-            // Pick a facility type.
-            uint32_t ft_idx = rng.next_uint(static_cast<uint32_t>(candidate_types.size()));
-            const FacilityType* ft = candidate_types[ft_idx];
-
-            // Pick a recipe that runs on this facility type.
-            auto ft_recipes = recipes.recipes_for_facility_type(ft->key);
-
-            // Filter by era.
-            std::vector<const Recipe*> available;
-            for (const auto* r : ft_recipes) {
-                if (r->era_available <= config.starting_era) {
-                    available.push_back(r);
-                }
-            }
-            if (available.empty())
-                continue;
-
-            uint32_t r_idx = rng.next_uint(static_cast<uint32_t>(available.size()));
-            const Recipe* recipe = available[r_idx];
-
-            Facility facility{};
-            facility.id = facility_id_counter++;
-            facility.business_id = biz.id;
-            facility.province_id = biz.province_id;
-            facility.recipe_id = recipe->id;
-            facility.tech_tier = recipe->min_tech_tier;
-            facility.output_rate_modifier = 1.0f;
-            facility.soil_health =
-                (ft->category == "agriculture") ? 0.8f + rng.next_float() * 0.2f : 1.0f;
-            facility.worker_count = std::min(
-                ft->max_workers,
-                static_cast<uint32_t>(3 + rng.next_uint(std::max(1u, ft->max_workers / 4))));
-            facility.is_operational = true;
-
-            world.facilities.push_back(std::move(facility));
-        }
-    }
-}
-
-// ===========================================================================
-// Technology seeding — assign baseline tech holdings to businesses
-// ===========================================================================
-
-void WorldGenerator::seed_technology(WorldState& world, DeterministicRNG& rng,
-                                     const TechnologyCatalog& tech_catalog,
-                                     const WorldGeneratorConfig& config) {
-    // Initialize GlobalTechnologyState.
-    world.technology.current_era = SimulationEra::era_1_turn_of_millennium;
-    world.technology.era_started_tick = 0;
-    world.technology.base_year = 2000;
-
-    // Seed initial domain knowledge from config defaults.
-    TechnologyConfig tech_config;
-    for (uint8_t i = 0; i < RESEARCH_DOMAIN_COUNT; ++i) {
-        world.technology.domain_knowledge[i] = tech_config.era1_domain_knowledge[i];
-    }
-
-    // Get baseline technology nodes (available at game start).
-    auto baseline = tech_catalog.baseline_nodes();
-    if (baseline.empty()) return;
-
-    // Sector-to-relevant-domain mapping for seeding.
-    // Businesses get baseline tech holdings in their relevant domains.
-    auto domains_for_sector = [](BusinessSector sector) -> std::vector<std::string> {
-        switch (sector) {
-            case BusinessSector::manufacturing:
-                return {"mechanical_engineering", "materials_science"};
-            case BusinessSector::food_beverage:
-                return {"biotechnology", "chemical_synthesis"};
-            case BusinessSector::agriculture:
-                return {"biotechnology"};
-            case BusinessSector::energy:
-                return {"energy_systems"};
-            case BusinessSector::technology:
-                return {"semiconductor_physics", "software_systems"};
-            case BusinessSector::research:
-                return {"semiconductor_physics", "software_systems", "biotechnology"};
-            case BusinessSector::transport_logistics:
-                return {"mechanical_engineering"};
-            case BusinessSector::criminal:
-                return {"illicit_chemistry", "chemical_synthesis"};
-            default:
-                return {"materials_science"};
-        }
-    };
-
-    for (auto& biz : world.npc_businesses) {
-        auto relevant_domains = domains_for_sector(biz.sector);
-
-        // Give each business holdings in baseline nodes matching their sector domains.
-        for (const TechnologyNode* node : baseline) {
-            bool relevant = false;
-            for (const auto& domain : relevant_domains) {
-                if (node->domain == domain) {
-                    relevant = true;
-                    break;
-                }
-            }
-            if (!relevant) continue;
-
-            TechHolding holding;
-            holding.node_key = node->node_key;
-            holding.holder_id = biz.id;
-            holding.stage = TechStage::commercialized;  // baseline tech is fully commercialized
-            holding.researched_tick = 0;
-            holding.commercialized_tick = 0;
-            holding.has_patent = false;
-            holding.internal_use_only = false;
-
-            // Maturation: baseline tech is mature (0.7-0.95 range, randomized).
-            float ceiling = tech_catalog.ceiling_for(node->node_key, config.starting_era);
-            if (ceiling < 0.0f) ceiling = 1.0f;  // fallback if no ceiling data
-            holding.maturation_ceiling = ceiling;
-            holding.maturation_level =
-                std::min(ceiling, 0.7f + rng.next_float() * 0.25f);
-
-            biz.actor_tech_state.holdings[node->node_key] = std::move(holding);
-        }
-
-        // Also give a few non-baseline Era 1 techs to a fraction of businesses
-        // (simulating early R&D investment that existed before year 2000).
-        auto era1_nodes = tech_catalog.nodes_available_at(1);
-        for (const TechnologyNode* node : era1_nodes) {
-            if (node->is_baseline) continue;
-
-            // Check domain relevance.
-            bool relevant = false;
-            for (const auto& domain : relevant_domains) {
-                if (node->domain == domain) {
-                    relevant = true;
-                    break;
-                }
-            }
-            if (!relevant) continue;
-
-            // ~20% chance of having started early research.
-            if (rng.next_float() > 0.20f) continue;
-
-            TechHolding holding;
-            holding.node_key = node->node_key;
-            holding.holder_id = biz.id;
-            holding.stage = TechStage::researched;  // early research, not yet commercialized
-            holding.researched_tick = 0;
-            holding.commercialized_tick = 0;
-            holding.has_patent = (node->patentable && rng.next_float() < 0.3f);
-            holding.internal_use_only = false;
-
-            float ceiling = tech_catalog.ceiling_for(node->node_key, config.starting_era);
-            if (ceiling < 0.0f) ceiling = 0.3f;  // low ceiling for early-era tech
-            holding.maturation_ceiling = ceiling;
-            // Early research: low maturation (0.05-0.25).
-            holding.maturation_level =
-                std::min(ceiling, 0.05f + rng.next_float() * 0.20f);
-
-            biz.actor_tech_state.holdings[node->node_key] = std::move(holding);
         }
     }
 }
@@ -3634,897 +3114,6 @@ void WorldGenerator::detect_special_features(WorldState& world, DeterministicRNG
 }
 
 // ===========================================================================
-// Stage 9 — Population and Infrastructure Seeding (WorldGen v0.18 §9)
-// ===========================================================================
-
-void WorldGenerator::seed_population_attractiveness(WorldState& world, DeterministicRNG& rng,
-                                                    const WorldGeneratorConfig& config) {
-    const auto& p = config.population;
-
-    for (auto& prov : world.provinces) {
-        // -----------------------------------------------------------------
-        // §9.2 — Disease burden from climate zone + standing water + elevation.
-        // -----------------------------------------------------------------
-        float burden = 0.0f;
-        KoppenZone kz = prov.climate.koppen_zone;
-
-        if (kz == KoppenZone::Af || kz == KoppenZone::Am) {
-            burden += 0.55f;  // hyperendemic tropical rainforest
-        } else if (kz == KoppenZone::Aw || kz == KoppenZone::BSh) {
-            burden += 0.35f;  // tropical savanna/steppe; seasonal
-        } else if (kz == KoppenZone::Cfa || kz == KoppenZone::Cwa) {
-            burden += 0.15f;  // humid subtropical
-        }
-
-        // Standing water amplifies vector habitat.
-        if (prov.climate.flood_vulnerability > 0.6f) {
-            burden += 0.15f;
-        }
-
-        // Elevation reduces disease (malaria vector altitude limit ~2,000m).
-        if (prov.geography.elevation_avg_m > 2000.0f) {
-            burden *= 0.10f;
-        } else if (prov.geography.elevation_avg_m > 1500.0f) {
-            burden *= 0.35f;
-        } else if (prov.geography.elevation_avg_m > 1000.0f) {
-            burden *= 0.65f;
-        }
-
-        prov.disease_burden = std::min(1.0f, std::max(0.0f, burden));
-
-        // -----------------------------------------------------------------
-        // §9.1 — Settlement attractiveness formula.
-        // -----------------------------------------------------------------
-
-        // Check for geothermal deposit.
-        bool has_geothermal = false;
-        for (const auto& d : prov.deposits) {
-            if (d.type == ResourceType::Geothermal) {
-                has_geothermal = true;
-                break;
-            }
-        }
-
-        float base =
-            prov.agricultural_productivity                * p.w_ag_productivity
-          + prov.geography.port_capacity                  * p.w_port_capacity
-          + prov.geography.river_access                   * p.w_river_access
-          + (1.0f - prov.geography.terrain_roughness)     * p.w_terrain_flatness
-          + (prov.soil_type == SoilType::Alluvial ? 1.0f : 0.0f) * p.w_alluvial_soil
-          + (has_geothermal ? 1.0f : 0.0f)                * p.w_geothermal
-          + (prov.soil_type == SoilType::Andisol ? 1.0f : 0.0f) * p.w_volcanic_soil;
-
-        // Altitude ceiling — hard physiological limits.
-        if (prov.geography.elevation_avg_m > 4500.0f) {
-            base *= p.alt_4500m_mult;
-        } else if (prov.geography.elevation_avg_m > 3500.0f) {
-            base *= p.alt_3500m_mult;
-        } else if (prov.geography.elevation_avg_m > 2500.0f) {
-            base *= p.alt_2500m_mult;
-        } else if (prov.geography.elevation_avg_m > 1500.0f) {
-            base *= p.alt_1500m_mult;
-        }
-
-        // Disease burden penalty (§9.2).
-        base *= (1.0f - prov.disease_burden * p.disease_max_penalty);
-
-        // Environmental penalties — genuine uninhabitability.
-        if (kz == KoppenZone::BWh || kz == KoppenZone::BWk) {
-            base *= p.desert_mult;
-        }
-        if (kz == KoppenZone::EF) {
-            base *= p.ice_cap_mult;
-        }
-        if (kz == KoppenZone::ET) {
-            base *= p.tundra_mult;
-        }
-        if (prov.geography.terrain_roughness > 0.85f) {
-            base *= p.extreme_terrain_mult;
-        }
-        if (prov.island_isolation) {
-            base *= p.island_isolation_mult;
-        }
-        if (prov.has_permafrost) {
-            base *= p.continuous_permafrost_mult;
-        }
-
-        float score = std::max(0.0f, std::min(1.0f, base));
-        prov.settlement_attractiveness = score;
-
-        // -----------------------------------------------------------------
-        // §9.3 — Infrastructure derivation from attractiveness.
-        // Flood vulnerability reduces infrastructure (costlier construction)
-        // without reducing population attractiveness.
-        // -----------------------------------------------------------------
-        float infra = score * p.infra_attract_scale
-                    - prov.climate.flood_vulnerability * p.infra_flood_penalty
-                    + (rng.next_float() - 0.5f) * 2.0f * p.infra_variance_sigma;
-        infra = std::max(0.0f, std::min(1.0f, infra));
-
-        // Blend with archetype infrastructure (50/50) to maintain diversity.
-        prov.infrastructure_rating = prov.infrastructure_rating * 0.50f + infra * 0.50f;
-
-        // -----------------------------------------------------------------
-        // §9.4 — Population adjustment from attractiveness.
-        // -----------------------------------------------------------------
-        float multiplier = p.multiplier_base + score * p.multiplier_range;
-        multiplier *= (p.rng_variation_base + rng.next_float() * p.rng_variation_range);
-
-        uint32_t new_pop = static_cast<uint32_t>(
-            static_cast<float>(prov.demographics.total_population) * multiplier + 0.5f);
-        prov.demographics.total_population = std::max(p.population_floor, new_pop);
-    }
-}
-
-// ===========================================================================
-// Stage 9.5 — Nation formation (WorldGen v0.18 §9.5)
-// ===========================================================================
-
-// Helper: compute terrain resistance for crossing a province link.
-// Mountains, rivers, and maritime crossings raise resistance; open flatland is cheap.
-// Uses NationFormationParams for all thresholds so the algorithm is tunable.
-static float compute_terrain_resistance(
-        const ProvinceLink& link, const Province& neighbor,
-        const WorldGeneratorConfig::NationFormationParams& nfp) {
-    float resistance = 1.0f;
-
-    // Maritime crossing: ocean is slower than land conquest
-    if (link.type == LinkType::Maritime) {
-        resistance += nfp.maritime_resistance;
-    }
-
-    // Elevation change proxy: high transit_terrain_cost means steep terrain
-    if (link.transit_terrain_cost > nfp.steep_terrain_threshold) {
-        resistance *= (1.0f + link.transit_terrain_cost);
-    }
-
-    // River crossing barrier (river links represent major navigable rivers)
-    if (link.type == LinkType::River) {
-        resistance *= nfp.river_crossing_mult;
-    }
-
-    // Uninhabitable provinces are expensive to cross (deserts, ice caps)
-    if (neighbor.settlement_attractiveness < 0.05f) {
-        resistance *= nfp.uninhabitable_mult;
-    }
-
-    return resistance;
-}
-
-// Helper: classify nation size from province count.
-static NationSize classify_nation_size(size_t province_count) {
-    if (province_count <= 3) return NationSize::Microstate;
-    if (province_count <= 12) return NationSize::Small;
-    if (province_count <= 40) return NationSize::Medium;
-    if (province_count <= 120) return NationSize::Large;
-    return NationSize::Continental;
-}
-
-// Helper: NationSize to string for JSON serialization.
-static const char* nation_size_str(NationSize s) {
-    switch (s) {
-        case NationSize::Microstate:   return "microstate";
-        case NationSize::Small:        return "small";
-        case NationSize::Medium:       return "medium";
-        case NationSize::Large:        return "large";
-        case NationSize::Continental:  return "continental";
-    }
-    return "small";
-}
-
-// Language families for V1 world generation. On non-Earth worlds all families
-// get equal geographic affinity weight — the list provides naming diversity.
-static const char* v1_language_families[] = {
-    "germanic", "romance", "slavic", "sinitic", "arabic",
-    "turkic",   "indic",   "bantu",  "austronesian", "quechuan",
-};
-static constexpr size_t v1_language_family_count = 10;
-
-void WorldGenerator::form_nations(WorldState& world, DeterministicRNG& rng,
-                                  const WorldGeneratorConfig& config) {
-    const auto& provinces = world.provinces;
-    const uint32_t prov_count = static_cast<uint32_t>(provinces.size());
-    if (prov_count == 0) return;
-
-    const auto& nfp = config.nation_formation;
-
-    // -----------------------------------------------------------------------
-    // Build h3_index → province_id lookup (O(1) neighbor resolution).
-    // Built FIRST so all subsequent passes use O(1) lookups instead of O(n) scans.
-    // -----------------------------------------------------------------------
-    std::unordered_map<H3Index, uint32_t> h3_to_idx;
-    h3_to_idx.reserve(prov_count);
-    for (uint32_t i = 0; i < prov_count; ++i) {
-        h3_to_idx[provinces[i].h3_index] = i;
-    }
-
-    // -----------------------------------------------------------------------
-    // §9.5.1 — Nation Seed Placement
-    // -----------------------------------------------------------------------
-    // Collect habitable provinces: attractiveness > 0.10, not EF ice cap.
-    // Provinces below uninhabitable_threshold (0.02) are excluded from nation
-    // assignment entirely — they become unclaimed territory per spec.
-    std::vector<uint32_t> habitable;
-    habitable.reserve(prov_count);
-    for (uint32_t i = 0; i < prov_count; ++i) {
-        const auto& p = provinces[i];
-        if (p.settlement_attractiveness > 0.10f &&
-            p.climate.koppen_zone != KoppenZone::EF) {
-            habitable.push_back(i);
-        }
-    }
-    if (habitable.empty()) {
-        // Degenerate world: all provinces are habitable for nation assignment.
-        for (uint32_t i = 0; i < prov_count; ++i) habitable.push_back(i);
-    }
-
-    // Target nation count: sqrt(habitable) × scale, clamped [min, max].
-    // For small worlds (habitable < min²/scale²), we gracefully reduce below
-    // the spec minimum to avoid more nations than provinces. The spec [20,400]
-    // targets Earth-scale (1000+ provinces); V1 with 6 provinces gets ~4.
-    uint32_t raw_target = static_cast<uint32_t>(
-        std::sqrt(static_cast<float>(habitable.size())) * nfp.seed_count_scale);
-    uint32_t target_count = std::clamp(raw_target, nfp.seed_count_min, nfp.seed_count_max);
-    // Never more nations than habitable provinces; minimum 2 for geopolitical tension.
-    target_count = std::min(target_count, static_cast<uint32_t>(habitable.size()));
-    target_count = std::max(target_count, std::min(2u, static_cast<uint32_t>(habitable.size())));
-
-    // Build attractiveness² weights for seed selection bias.
-    // Attractiveness² biases toward high-value inland/coastal cores, not thin margins.
-    std::vector<float> weights(habitable.size());
-    float total_weight = 0.0f;
-    for (size_t i = 0; i < habitable.size(); ++i) {
-        float a = provinces[habitable[i]].settlement_attractiveness;
-        weights[i] = a * a;
-        total_weight += weights[i];
-    }
-
-    // Track which provinces are available for seed selection.
-    // Using a bool vector is O(1) per check vs O(log n) for unordered_set.
-    std::vector<bool> available(prov_count, false);
-    for (uint32_t pid : habitable) available[pid] = true;
-
-    std::vector<uint32_t> seed_province_ids;
-    seed_province_ids.reserve(target_count);
-
-    // Minimum separation in graph hops (BFS distance through ProvinceLinks).
-    // Spec §9.5.1: "no two seeds closer than 3 H3 grid-disks."
-    // For small worlds (< seed_separation * 2 provinces), reduce to avoid
-    // excluding all candidates.
-    uint32_t effective_separation = nfp.seed_separation;
-    if (habitable.size() < effective_separation * 2) {
-        effective_separation = 0;  // small worlds: no separation constraint
-    }
-
-    // Weighted sampling loop. On each iteration we pick one seed, then
-    // exclude neighbors within effective_separation via BFS.
-    // Instead of recomputing total_weight from scratch each time (O(n)),
-    // we subtract removed weights incrementally.
-    uint32_t max_attempts = static_cast<uint32_t>(habitable.size()) * 3;
-    for (uint32_t attempt = 0;
-         seed_province_ids.size() < target_count && attempt < max_attempts;
-         ++attempt) {
-        if (total_weight <= 0.0f) break;
-
-        // Weighted random selection.
-        float roll = rng.next_float() * total_weight;
-        float cumulative = 0.0f;
-        uint32_t chosen = habitable[0];
-        size_t chosen_idx = 0;
-        for (size_t i = 0; i < habitable.size(); ++i) {
-            if (!available[habitable[i]]) continue;
-            cumulative += weights[i];
-            if (cumulative >= roll) {
-                chosen = habitable[i];
-                chosen_idx = i;
-                break;
-            }
-        }
-
-        if (!available[chosen]) continue;
-
-        seed_province_ids.push_back(chosen);
-        available[chosen] = false;
-        total_weight -= weights[chosen_idx];
-        weights[chosen_idx] = 0.0f;
-
-        // BFS exclusion zone: mark neighbors within effective_separation as unavailable.
-        if (effective_separation > 0) {
-            std::queue<std::pair<uint32_t, uint32_t>> bfs;
-            bfs.push({chosen, 0});
-            std::unordered_set<uint32_t> visited_bfs;
-            visited_bfs.insert(chosen);
-            while (!bfs.empty()) {
-                auto [cur, dist] = bfs.front();
-                bfs.pop();
-                if (dist >= effective_separation) continue;
-                for (const auto& link : provinces[cur].links) {
-                    auto it = h3_to_idx.find(link.neighbor_h3);
-                    if (it == h3_to_idx.end()) continue;
-                    uint32_t nid = it->second;
-                    if (visited_bfs.count(nid)) continue;
-                    visited_bfs.insert(nid);
-                    if (available[nid]) {
-                        available[nid] = false;
-                        // Find and zero this province's weight to keep total_weight accurate.
-                        for (size_t i = 0; i < habitable.size(); ++i) {
-                            if (habitable[i] == nid) {
-                                total_weight -= weights[i];
-                                weights[i] = 0.0f;
-                                break;
-                            }
-                        }
-                    }
-                    bfs.push({nid, dist + 1});
-                }
-            }
-        }
-    }
-
-    // Fallback: if no seeds were placed, use the highest-attractiveness province.
-    if (seed_province_ids.empty()) {
-        uint32_t best = 0;
-        for (uint32_t i = 1; i < prov_count; ++i) {
-            if (provinces[i].settlement_attractiveness >
-                provinces[best].settlement_attractiveness) {
-                best = i;
-            }
-        }
-        seed_province_ids.push_back(best);
-    }
-
-    const uint32_t nation_count = static_cast<uint32_t>(seed_province_ids.size());
-
-    // -----------------------------------------------------------------------
-    // §9.5.2 — Voronoi Growth with Terrain Resistance
-    // -----------------------------------------------------------------------
-    // Identify uninhabitable provinces that should remain unclaimed.
-    std::vector<bool> uninhabitable(prov_count, false);
-    for (uint32_t i = 0; i < prov_count; ++i) {
-        if (provinces[i].settlement_attractiveness < nfp.uninhabitable_threshold &&
-            provinces[i].climate.koppen_zone == KoppenZone::EF) {
-            uninhabitable[i] = true;
-        }
-        // Deep desert with near-zero attractiveness also unclaimed.
-        if (provinces[i].settlement_attractiveness < nfp.uninhabitable_threshold &&
-            (provinces[i].climate.koppen_zone == KoppenZone::BWh ||
-             provinces[i].climate.koppen_zone == KoppenZone::BWk)) {
-            uninhabitable[i] = true;
-        }
-    }
-
-    // Identify island provinces (all links are Maritime) for special post-pass assignment.
-    std::vector<bool> is_island(prov_count, false);
-    for (uint32_t i = 0; i < prov_count; ++i) {
-        if (provinces[i].links.empty()) continue;
-        bool all_maritime = true;
-        for (const auto& link : provinces[i].links) {
-            if (link.type != LinkType::Maritime) {
-                all_maritime = false;
-                break;
-            }
-        }
-        if (all_maritime) is_island[i] = true;
-    }
-
-    // Dijkstra-style priority queue: (cost, province_id, nation_index).
-    using PQEntry = std::tuple<float, uint32_t, uint32_t>;
-    std::priority_queue<PQEntry, std::vector<PQEntry>, std::greater<PQEntry>> pq;
-
-    std::vector<float> best_cost(prov_count, std::numeric_limits<float>::max());
-    std::vector<int32_t> assignment(prov_count, -1);  // -1 = unassigned
-
-    // Initialize seeds. Uninhabitable seeds shouldn't occur (filtered in habitable),
-    // but guard anyway.
-    for (uint32_t ni = 0; ni < nation_count; ++ni) {
-        uint32_t pid = seed_province_ids[ni];
-        best_cost[pid] = 0.0f;
-        assignment[pid] = static_cast<int32_t>(ni);
-        pq.push({0.0f, pid, ni});
-    }
-
-    // Main Voronoi flood-fill. Skip uninhabitable and island provinces (islands
-    // get special assignment in the post-pass below).
-    while (!pq.empty()) {
-        auto [cost, current, nation_idx] = pq.top();
-        pq.pop();
-
-        if (cost > best_cost[current]) continue;
-
-        for (const auto& link : provinces[current].links) {
-            auto it = h3_to_idx.find(link.neighbor_h3);
-            if (it == h3_to_idx.end()) continue;
-            uint32_t neighbor_id = it->second;
-
-            // Skip uninhabitable provinces — they remain unclaimed territory.
-            if (uninhabitable[neighbor_id]) continue;
-
-            // Skip islands during main pass — they get post-pass assignment.
-            if (is_island[neighbor_id]) continue;
-
-            if (assignment[neighbor_id] >= 0 &&
-                best_cost[neighbor_id] <= cost) continue;
-
-            float resistance = compute_terrain_resistance(link, provinces[neighbor_id], nfp);
-            float new_cost = cost + resistance;
-
-            if (new_cost < best_cost[neighbor_id]) {
-                best_cost[neighbor_id] = new_cost;
-                assignment[neighbor_id] = static_cast<int32_t>(nation_idx);
-                pq.push({new_cost, neighbor_id, nation_idx});
-            }
-        }
-    }
-
-    // §9.5.2 post-pass: assign island provinces to the nation that owns the
-    // maritime-adjacent province with the longest shared border.
-    for (uint32_t i = 0; i < prov_count; ++i) {
-        if (!is_island[i]) continue;
-        if (uninhabitable[i]) continue;
-
-        int32_t best_nation = -1;
-        float best_border_km = -1.0f;
-        for (const auto& link : provinces[i].links) {
-            auto it = h3_to_idx.find(link.neighbor_h3);
-            if (it == h3_to_idx.end()) continue;
-            uint32_t nid = it->second;
-            if (assignment[nid] >= 0 && link.shared_border_km > best_border_km) {
-                best_border_km = link.shared_border_km;
-                best_nation = assignment[nid];
-            }
-        }
-        // If no assigned neighbor found (all neighbors are also islands),
-        // try any assigned neighbor via BFS at distance 2.
-        if (best_nation < 0) {
-            for (const auto& link : provinces[i].links) {
-                auto it = h3_to_idx.find(link.neighbor_h3);
-                if (it == h3_to_idx.end()) continue;
-                uint32_t nid = it->second;
-                for (const auto& link2 : provinces[nid].links) {
-                    auto it2 = h3_to_idx.find(link2.neighbor_h3);
-                    if (it2 == h3_to_idx.end()) continue;
-                    if (assignment[it2->second] >= 0) {
-                        best_nation = assignment[it2->second];
-                        break;
-                    }
-                }
-                if (best_nation >= 0) break;
-            }
-        }
-        if (best_nation >= 0) {
-            assignment[i] = best_nation;
-        }
-        // else: truly isolated island — remains unclaimed (rare).
-    }
-
-    // Provinces that remain unassigned (no links, not uninhabitable) → first nation.
-    for (uint32_t i = 0; i < prov_count; ++i) {
-        if (assignment[i] < 0 && !uninhabitable[i]) {
-            assignment[i] = 0;
-        }
-    }
-
-    // -----------------------------------------------------------------------
-    // Build Nation objects
-    // -----------------------------------------------------------------------
-    world.nations.clear();
-    world.region_groups.clear();
-
-    // Scalable name generation: combine prefix + root, cycling through roots.
-    // 8 prefixes × 60 roots = 480 unique names, sufficient for max 400 nations.
-    static const char* nation_prefixes[] = {
-        "Republic of ", "Kingdom of ", "Federation of ", "Commonwealth of ",
-        "State of ", "Dominion of ", "Principality of ", "Union of ",
-    };
-    static constexpr size_t n_prefixes = 8;
-    static const char* nation_roots[] = {
-        "Avalon",   "Corvana",  "Delvoria", "Estmarch",  "Fenwick",
-        "Galdria",  "Haldane",  "Irindel",  "Jastova",   "Keldara",
-        "Lorantia", "Morvaine", "Narthia",  "Ostenveld",  "Pelluria",
-        "Quentara", "Rhedania", "Sylvarna", "Torvalis",   "Ulmendia",
-        "Veldmark", "Wyverna",  "Xandria",  "Yelthar",   "Zaranthia",
-        "Almuria",  "Brestova", "Calendra", "Damavand",  "Eldrathi",
-        "Frostmark","Gelvaine", "Halvoria", "Istharan",  "Jorvald",
-        "Korinth",  "Lynthia",  "Marvesta", "Novogard",  "Orelund",
-        "Praethos", "Quelaria", "Rochfort", "Silvaine",  "Thalmund",
-        "Umbravia", "Valdanis", "Westmark", "Xenthira",  "Yrvandia",
-        "Zephyria", "Arcandia", "Belmoris", "Calderon",  "Dravenholm",
-        "Elysar",   "Fyrnhold", "Grenmark", "Helmgaard", "Isolvar",
-    };
-    static constexpr size_t n_roots = 60;
-
-    for (uint32_t ni = 0; ni < nation_count; ++ni) {
-        Nation nation{};
-        nation.id = ni;
-
-        // Deterministic name from RNG + cycling through root/prefix pools.
-        uint32_t prefix_idx = static_cast<uint32_t>(rng.next_float() * n_prefixes) % n_prefixes;
-        uint32_t root_idx = ni % n_roots;
-        nation.name = std::string(nation_prefixes[prefix_idx]) + nation_roots[root_idx];
-        // Currency code: first 3 chars of root, uppercased.
-        std::string root_str(nation_roots[root_idx]);
-        nation.currency_code = root_str.substr(0, 3);
-        for (auto& c : nation.currency_code) c = static_cast<char>(std::toupper(c));
-
-        // Government type weighted by RNG roll.
-        float gov_roll = rng.next_float();
-        if (gov_roll < 0.50f) nation.government_type = GovernmentType::Democracy;
-        else if (gov_roll < 0.75f) nation.government_type = GovernmentType::Federation;
-        else if (gov_roll < 0.95f) nation.government_type = GovernmentType::Autocracy;
-        else nation.government_type = GovernmentType::FailedState;
-
-        nation.political_cycle = {0, 0.50f + rng.next_float() * 0.20f, false,
-                                   365 * (3 + static_cast<uint32_t>(rng.next_float() * 3))};
-        nation.corporate_tax_rate = 0.15f + rng.next_float() * 0.20f;
-        nation.income_tax_rate_top_bracket = 0.20f + rng.next_float() * 0.30f;
-        nation.trade_balance_fraction = 0.0f;
-        nation.inflation_rate = 0.01f + rng.next_float() * 0.05f;
-        nation.credit_rating = 0.40f + rng.next_float() * 0.50f;
-        nation.tariff_schedule = nullptr;
-
-        // First nation (player's home) is LOD 0; others are LOD 1.
-        if (ni == 0) {
-            nation.lod1_profile = std::nullopt;
-        } else {
-            Lod1NationProfile lod1{};
-            lod1.export_margin = 0.10f + rng.next_float() * 0.15f;
-            lod1.import_premium = 0.05f + rng.next_float() * 0.10f;
-            lod1.trade_openness = 0.30f + rng.next_float() * 0.40f;
-            lod1.tech_tier_modifier = 0.80f + rng.next_float() * 0.40f;
-            lod1.population_modifier = 0.70f + rng.next_float() * 0.60f;
-            lod1.research_investment = 0.0f;
-            lod1.current_tier = 1;
-            nation.lod1_profile = lod1;
-        }
-
-        // Collect member provinces.
-        for (uint32_t pi = 0; pi < prov_count; ++pi) {
-            if (assignment[pi] == static_cast<int32_t>(ni)) {
-                nation.province_ids.push_back(pi);
-            }
-        }
-
-        nation.size_class = classify_nation_size(nation.province_ids.size());
-
-        world.nations.push_back(std::move(nation));
-    }
-
-    // Assign nation_id on each province. Unclaimed provinces get a sentinel
-    // value equal to nation_count (no valid nation). For V1 runtime compat,
-    // unclaimed provinces still get nation_id = 0 (player's nation handles them).
-    for (uint32_t i = 0; i < prov_count; ++i) {
-        if (assignment[i] >= 0) {
-            world.provinces[i].nation_id = static_cast<uint32_t>(assignment[i]);
-        } else {
-            // Unclaimed territory: assign to nation 0 for V1 runtime compat.
-            // Stage 10 can read uninhabitable[] to detect these.
-            world.provinces[i].nation_id = 0;
-        }
-    }
-
-    // Create one region per province (V1 simplification).
-    for (uint32_t r = 0; r < prov_count; ++r) {
-        Region region{};
-        region.id = r;
-        region.fictional_name = "Region_" + std::to_string(r);
-        region.nation_id = world.provinces[r].nation_id;
-        region.province_ids.push_back(r);
-        world.region_groups.push_back(std::move(region));
-    }
-
-    // -----------------------------------------------------------------------
-    // §9.5.3 — Language Family Assignment
-    // -----------------------------------------------------------------------
-    // Two-pass: initial geographic assignment, then neighbor propagation.
-
-    // Pass 1: each nation's seed province gets a language family by geographic affinity.
-    for (auto& nation : world.nations) {
-        if (nation.province_ids.empty()) continue;
-        // Use seed province (first province_id is the seed from Voronoi growth).
-        uint32_t seed_pid = seed_province_ids[std::min(
-            nation.id, static_cast<uint32_t>(seed_province_ids.size() - 1))];
-        const auto& seed_prov = provinces[seed_pid];
-        float abs_lat = std::fabs(seed_prov.geography.latitude);
-
-        // Build affinity weights for each language family.
-        float family_weights[v1_language_family_count];
-        for (size_t fi = 0; fi < v1_language_family_count; ++fi) {
-            family_weights[fi] = 1.0f;  // neutral default
-        }
-        // Germanic: mid-high latitude, wetter regions
-        if (abs_lat > 45.0f && abs_lat < 65.0f) family_weights[0] = 1.5f;
-        // Romance: mid latitude
-        if (abs_lat > 25.0f && abs_lat < 50.0f) family_weights[1] = 1.4f;
-        // Slavic: continental mid-high latitude
-        if (abs_lat > 40.0f && abs_lat < 65.0f) family_weights[2] = 1.3f;
-        // Sinitic: mid latitude East Asian analog
-        if (abs_lat > 20.0f && abs_lat < 45.0f) family_weights[3] = 1.3f;
-        // Arabic: low-mid latitude, arid zones
-        if (abs_lat < 35.0f) family_weights[4] = 1.4f;
-        // Turkic: continental steppe
-        if (abs_lat > 30.0f && abs_lat < 55.0f) family_weights[5] = 1.2f;
-        // Indic: tropical-subtropical
-        if (abs_lat < 30.0f) family_weights[6] = 1.3f;
-        // Bantu: tropical Africa analog
-        if (abs_lat < 25.0f) family_weights[7] = 1.4f;
-        // Austronesian: coastal tropical (island/maritime cultures)
-        if (abs_lat < 20.0f && seed_prov.geography.coastal_length_km > 20.0f)
-            family_weights[8] = 1.5f;
-        // Quechuan: highland
-        if (seed_prov.geography.elevation_avg_m > 2000.0f) family_weights[9] = 1.5f;
-
-        float total = 0.0f;
-        for (size_t fi = 0; fi < v1_language_family_count; ++fi) total += family_weights[fi];
-        float roll = rng.next_float() * total;
-        float cum = 0.0f;
-        size_t chosen_fi = 0;
-        for (size_t fi = 0; fi < v1_language_family_count; ++fi) {
-            cum += family_weights[fi];
-            if (cum >= roll) {
-                chosen_fi = fi;
-                break;
-            }
-        }
-        nation.language_family_id = v1_language_families[chosen_fi];
-    }
-
-    // Pass 2: neighbor propagation — chance to inherit largest neighbor's language.
-    // Process largest nations first (they set regional tone).
-    std::vector<uint32_t> nation_order(nation_count);
-    for (uint32_t ni = 0; ni < nation_count; ++ni) nation_order[ni] = ni;
-    std::sort(nation_order.begin(), nation_order.end(),
-              [&](uint32_t a, uint32_t b) {
-                  return world.nations[a].province_ids.size() >
-                         world.nations[b].province_ids.size();
-              });
-
-    // Build nation adjacency from province links.
-    std::unordered_map<uint32_t, std::unordered_set<uint32_t>> nation_adj;
-    for (uint32_t pi = 0; pi < prov_count; ++pi) {
-        if (assignment[pi] < 0) continue;  // skip unclaimed
-        for (const auto& link : provinces[pi].links) {
-            auto it = h3_to_idx.find(link.neighbor_h3);
-            if (it == h3_to_idx.end()) continue;
-            if (assignment[it->second] < 0) continue;  // skip unclaimed neighbor
-            uint32_t ni_a = world.provinces[pi].nation_id;
-            uint32_t ni_b = world.provinces[it->second].nation_id;
-            if (ni_a != ni_b) {
-                nation_adj[ni_a].insert(ni_b);
-                nation_adj[ni_b].insert(ni_a);
-            }
-        }
-    }
-
-    for (uint32_t ni : nation_order) {
-        auto adj_it = nation_adj.find(ni);
-        if (adj_it == nation_adj.end() || adj_it->second.empty()) continue;
-
-        // Find largest neighbor.
-        uint32_t largest_neighbor = *adj_it->second.begin();
-        for (uint32_t nj : adj_it->second) {
-            if (world.nations[nj].province_ids.size() >
-                world.nations[largest_neighbor].province_ids.size()) {
-                largest_neighbor = nj;
-            }
-        }
-
-        if (rng.next_float() < nfp.language_propagation_chance) {
-            world.nations[ni].language_family_id =
-                world.nations[largest_neighbor].language_family_id;
-        }
-
-        // Secondary language: most common different-language neighbor.
-        std::unordered_map<std::string, uint32_t> diff_lang_counts;
-        for (uint32_t nj : adj_it->second) {
-            if (world.nations[nj].language_family_id !=
-                world.nations[ni].language_family_id) {
-                diff_lang_counts[world.nations[nj].language_family_id]++;
-            }
-        }
-        if (!diff_lang_counts.empty()) {
-            auto best = std::max_element(
-                diff_lang_counts.begin(), diff_lang_counts.end(),
-                [](const auto& a, const auto& b) { return a.second < b.second; });
-            world.nations[ni].secondary_language_id = best->first;
-        }
-    }
-
-    // -----------------------------------------------------------------------
-    // §9.5.4 — Border Change Seeding
-    // -----------------------------------------------------------------------
-    // Build a notable resource threshold: 75th percentile of all deposit quantities.
-    std::vector<float> all_quantities;
-    for (const auto& prov : world.provinces) {
-        for (const auto& dep : prov.deposits) {
-            all_quantities.push_back(dep.quantity);
-        }
-    }
-    float notable_threshold = 0.50f;
-    if (!all_quantities.empty()) {
-        std::sort(all_quantities.begin(), all_quantities.end());
-        notable_threshold = all_quantities[all_quantities.size() * 3 / 4];
-    }
-
-    for (auto& prov : world.provinces) {
-        float instability = 0.0f;
-
-        // Border location: frontier provinces are contested.
-        bool is_border_prov = false;
-        for (const auto& link : prov.links) {
-            auto it = h3_to_idx.find(link.neighbor_h3);
-            if (it == h3_to_idx.end()) continue;
-            if (link.type == LinkType::Land || link.type == LinkType::River) {
-                if (world.provinces[it->second].nation_id != prov.nation_id) {
-                    is_border_prov = true;
-                    break;
-                }
-            }
-        }
-        if (is_border_prov) instability += nfp.border_instability;
-
-        // Strategic value: notable resources.
-        for (const auto& dep : prov.deposits) {
-            if (dep.quantity > notable_threshold) {
-                instability += nfp.resource_instability;
-                break;
-            }
-        }
-
-        // High settlement attractiveness attracts conquest.
-        if (prov.settlement_attractiveness > 0.70f) instability += nfp.attractiveness_instability;
-
-        // Mountain passes and plateaus are contested.
-        if (prov.is_mountain_pass) instability += nfp.chokepoint_instability;
-
-        // Infra gap: predicted vs actual infrastructure.
-        float predicted_infra = prov.settlement_attractiveness * 0.70f;
-        prov.infra_gap = prov.infrastructure_rating - predicted_infra;
-
-        // Colonial signature: infra_gap > 0.20 suggests external development.
-        if (prov.infra_gap > 0.20f) instability += nfp.colonial_instability;
-
-        // Poisson draw: instability → expected count → actual count.
-        float expected = instability * nfp.instability_to_expected;
-        int32_t count = 0;
-        float remaining = expected;
-        while (remaining > 0.0f && count < nfp.max_border_changes) {
-            float u = rng.next_float();
-            if (u < 0.001f) u = 0.001f;
-            remaining += std::log(u);  // log(u) is negative
-            if (remaining > 0.0f) ++count;
-        }
-        prov.border_change_count = std::max(0, std::min(count, nfp.max_border_changes));
-
-        // Colonial development flag.
-        prov.has_colonial_development_event = (
-            prov.infra_gap > 0.20f &&
-            prov.border_change_count > 0 &&
-            prov.infrastructure_rating > 0.50f);
-    }
-
-    // -----------------------------------------------------------------------
-    // Aggregate nation-level fields
-    // -----------------------------------------------------------------------
-    for (auto& nation : world.nations) {
-        if (nation.province_ids.empty()) continue;
-
-        float sum_attract = 0.0f;
-        float sum_infra = 0.0f;
-        for (uint32_t pid : nation.province_ids) {
-            sum_attract += world.provinces[pid].settlement_attractiveness;
-            sum_infra += world.provinces[pid].infrastructure_rating;
-        }
-        float n = static_cast<float>(nation.province_ids.size());
-        nation.gdp_index = std::clamp(sum_attract / n, 0.0f, 1.0f);
-        float mean_infra = sum_infra / n;
-        nation.governance_quality = std::clamp(
-            mean_infra + (rng.next_float() - 0.5f) * 0.20f, 0.0f, 1.0f);
-
-        for (uint32_t pid : nation.province_ids) {
-            if (world.provinces[pid].has_colonial_development_event) {
-                nation.is_colonial_power = true;
-                break;
-            }
-        }
-    }
-}
-
-// ===========================================================================
-// Stage 9.6 — Nomadic population (WorldGen v0.18 §9.6)
-// ===========================================================================
-
-void WorldGenerator::seed_nomadic_population(WorldState& world, DeterministicRNG& rng,
-                                              const WorldGeneratorConfig& config) {
-    const float realisation = config.nation_formation.nomadic_realisation_factor;
-    for (auto& prov : world.provinces) {
-        float pastoral_cap = 0.0f;
-
-        // Pastoral carrying capacity from Köppen zone.
-        switch (prov.climate.koppen_zone) {
-            case KoppenZone::BSk:
-            case KoppenZone::BSh:
-                pastoral_cap = 0.65f;  // steppe: classic pastoral zone
-                break;
-            case KoppenZone::BWh:
-            case KoppenZone::BWk:
-                pastoral_cap = 0.20f;  // desert: sparse; oases anchor movement
-                break;
-            case KoppenZone::Aw:
-                pastoral_cap = 0.45f;  // savanna: seasonal transhumance
-                break;
-            case KoppenZone::ET:
-                pastoral_cap = 0.30f;  // tundra: reindeer pastoralism
-                break;
-            case KoppenZone::Dfc:
-            case KoppenZone::Dfd:
-                pastoral_cap = 0.15f;  // taiga fringe; limited
-                break;
-            default:
-                pastoral_cap = 0.0f;
-                break;
-        }
-
-        // Reduce by terrain roughness (mountains break up grazing range).
-        pastoral_cap *= (1.0f - prov.geography.terrain_roughness * 0.50f);
-
-        // Reduce where agricultural productivity already supports dense settlement.
-        pastoral_cap *= std::max(0.0f, 1.0f - prov.agricultural_productivity * 1.5f);
-
-        prov.pastoral_carrying_capacity = std::clamp(pastoral_cap, 0.0f, 1.0f);
-
-        // Nomadic fraction: highest where pastoral capacity is high and settled capacity low.
-        if (pastoral_cap > 0.10f) {
-            prov.nomadic_population_fraction = pastoral_cap * realisation;
-        } else {
-            prov.nomadic_population_fraction = 0.0f;
-        }
-    }
-}
-
-// ===========================================================================
-// Stage 9.7 — Nation capital seeding (WorldGen v0.18 §9.7)
-// ===========================================================================
-
-void WorldGenerator::seed_nation_capitals(WorldState& world,
-                                           const WorldGeneratorConfig& config) {
-    (void)config;  // available for future tuning
-    for (auto& nation : world.nations) {
-        if (nation.province_ids.empty()) continue;
-
-        // Filter eligible provinces: no continuous permafrost, elevation < 4000m,
-        // not badlands/tundra/glacial terrain.
-        std::vector<uint32_t> eligible;
-        for (uint32_t pid : nation.province_ids) {
-            const auto& p = world.provinces[pid];
-            if (p.has_permafrost &&
-                p.climate.koppen_zone == KoppenZone::EF) continue;
-            if (p.geography.elevation_avg_m >= 4000.0f) continue;
-            if (p.has_badlands) continue;
-            if (p.is_glacial_scoured &&
-                p.agricultural_productivity < 0.10f) continue;
-            eligible.push_back(pid);
-        }
-        if (eligible.empty()) {
-            eligible = nation.province_ids;  // fallback: no constraint
-        }
-
-        // Select highest settlement_attractiveness province.
-        uint32_t capital_pid = eligible[0];
-        float best_attract = world.provinces[capital_pid].settlement_attractiveness;
-        for (uint32_t pid : eligible) {
-            float a = world.provinces[pid].settlement_attractiveness;
-            if (a > best_attract) {
-                best_attract = a;
-                capital_pid = pid;
-            }
-        }
-
-        world.provinces[capital_pid].is_nation_capital = true;
-        nation.capital_province_id = capital_pid;
-    }
-}
-
-// ===========================================================================
 // Stage 10.0 — Province archetype classification (24-archetype taxonomy)
 // ===========================================================================
 
@@ -6184,6 +4773,17 @@ static const char* lod_str(SimulationLOD lod) {
         case SimulationLOD::statistical: return "statistical";
     }
     return "unknown";
+}
+
+static const char* nation_size_str(NationSize s) {
+    switch (s) {
+        case NationSize::Microstate:   return "microstate";
+        case NationSize::Small:        return "small";
+        case NationSize::Medium:       return "medium";
+        case NationSize::Large:        return "large";
+        case NationSize::Continental:  return "continental";
+    }
+    return "small";
 }
 
 static const char* government_type_str(GovernmentType gt) {
